# docs/static/openapi/v1.yaml
openapi: 3.1.0
info:
  title: imgsrv API
  version: v1
  description: Contract for the implemented imgsrv v1 upload and catalog API.
servers:
  - url: /
paths:
  /v1/uploads:
    post:
      operationId: beginUpload
      summary: Begin an upload session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BeginUploadRequest'
      responses:
        '201':
          description: Upload session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadSession'
        '400':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /v1/uploads/{upload_id}:
    get:
      operationId: getUpload
      summary: Get upload session status.
      parameters:
        - $ref: '#/components/parameters/UploadID'
      responses:
        '200':
          description: Current upload session state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadSession'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /v1/uploads/{upload_id}/parts/{part_number}:
    put:
      operationId: putUploadPart
      summary: Upload or replace one multipart upload part.
      parameters:
        - $ref: '#/components/parameters/UploadID'
        - name: part_number
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 10000
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Upload part accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadPart'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '412':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /v1/uploads/{upload_id}/complete:
    post:
      operationId: completeUpload
      summary: Complete an upload session.
      parameters:
        - $ref: '#/components/parameters/UploadID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteUploadRequest'
      responses:
        '200':
          description: Upload session completed or already terminally advanced.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadSession'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '412':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /v1/uploads/{upload_id}/abort:
    post:
      operationId: abortUpload
      summary: Abort an upload session.
      parameters:
        - $ref: '#/components/parameters/UploadID'
      responses:
        '200':
          description: Upload session aborted or already aborted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadSession'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '412':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /v1/images:
    post:
      operationId: createImage
      summary: Create an image namespace.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateImageRequest'
      responses:
        '201':
          description: Image namespace created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '400':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /v1/images/{name}/versions:
    post:
      operationId: createDraftVersion
      summary: Create a draft image version.
      parameters:
        - $ref: '#/components/parameters/ImageName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDraftVersionRequest'
      responses:
        '201':
          description: Draft image version created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageVersion'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /v1/images/{name}/versions/{version}:
    get:
      operationId: getVersionManifest
      summary: Get an exact draft or published version manifest.
      parameters:
        - $ref: '#/components/parameters/ImageName'
        - $ref: '#/components/parameters/ImageVersion'
      responses:
        '200':
          description: Exact image version manifest.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Manifest'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /v1/images/{name}/versions/{version}/artifacts:
    post:
      operationId: addArtifact
      summary: Add a primary artifact to a draft version.
      parameters:
        - $ref: '#/components/parameters/ImageName'
        - $ref: '#/components/parameters/ImageVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddArtifactRequest'
      responses:
        '201':
          description: Artifact declaration created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artifact'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '412':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /v1/images/{name}/versions/{version}/artifacts/{artifact_id}/attachments:
    post:
      operationId: addAttachment
      summary: Add an attachment to a draft artifact.
      parameters:
        - $ref: '#/components/parameters/ImageName'
        - $ref: '#/components/parameters/ImageVersion'
        - $ref: '#/components/parameters/ArtifactID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAttachmentRequest'
      responses:
        '201':
          description: Attachment declaration created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attachment'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '412':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /v1/images/{name}/versions/{version}/publish:
    post:
      operationId: publishVersion
      summary: Publish a draft image version.
      parameters:
        - $ref: '#/components/parameters/ImageName'
        - $ref: '#/components/parameters/ImageVersion'
      responses:
        '200':
          description: Image version published.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageVersion'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '412':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
components:
  parameters:
    UploadID:
      name: upload_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UploadID'
    ImageName:
      name: name
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9][a-z0-9._-]{0,127}$'
    ImageVersion:
      name: version
      in: path
      required: true
      schema:
        type: string
        pattern: '^[A-Za-z0-9][A-Za-z0-9._+:-]{0,127}$'
    ArtifactID:
      name: artifact_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ArtifactID'
  responses:
    Problem:
      description: RFC 9457 problem details.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Digest:
      type: string
      pattern: '^sha256:[0-9a-f]{64}$'
      examples:
        - sha256:0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef
    UploadID:
      type: string
      format: uuid
    ArtifactID:
      type: string
      format: uuid
    UploadState:
      type: string
      enum:
        - created
        - uploading
        - completed
        - ingesting
        - ready
        - failed
        - aborted
    ImageVersionState:
      type: string
      enum:
        - draft
        - published
    ArtifactFormat:
      type: string
      enum:
        - raw
        - qcow2
    BeginUploadRequest:
      type: object
      required:
        - expected_digest
        - expected_size_bytes
      properties:
        expected_digest:
          $ref: '#/components/schemas/Digest'
        expected_size_bytes:
          type: integer
          format: int64
          minimum: 0
        media_type_hint:
          type: string
          minLength: 1
        filename_hint:
          type: string
          minLength: 1
      additionalProperties: false
    CompleteUploadRequest:
      type: object
      required:
        - parts
      properties:
        parts:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CompleteUploadPart'
      additionalProperties: false
    CompleteUploadPart:
      type: object
      required:
        - number
        - etag
        - size_bytes
      properties:
        number:
          type: integer
          minimum: 1
          maximum: 10000
        etag:
          type: string
          minLength: 1
        size_bytes:
          type: integer
          format: int64
          minimum: 0
      additionalProperties: false
    UploadSession:
      type: object
      required:
        - id
        - expected_digest
        - expected_size_bytes
        - state
        - expires_at
      properties:
        id:
          $ref: '#/components/schemas/UploadID'
        expected_digest:
          $ref: '#/components/schemas/Digest'
        expected_size_bytes:
          type: integer
          format: int64
          minimum: 0
        state:
          $ref: '#/components/schemas/UploadState'
        expires_at:
          type: string
          format: date-time
        media_type_hint:
          type: string
          minLength: 1
        filename_hint:
          type: string
          minLength: 1
      additionalProperties: false
    UploadPart:
      type: object
      required:
        - upload_id
        - part_number
        - etag
        - size_bytes
      properties:
        upload_id:
          $ref: '#/components/schemas/UploadID'
        part_number:
          type: integer
          minimum: 1
          maximum: 10000
        etag:
          type: string
          minLength: 1
        size_bytes:
          type: integer
          format: int64
          minimum: 0
      additionalProperties: false
    CreateImageRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          pattern: '^[a-z0-9][a-z0-9._-]{0,127}$'
        display_name:
          type: string
        description:
          type: string
      additionalProperties: false
    CreateDraftVersionRequest:
      type: object
      required:
        - version
      properties:
        version:
          type: string
          pattern: '^[A-Za-z0-9][A-Za-z0-9._+:-]{0,127}$'
      additionalProperties: false
    AddArtifactRequest:
      type: object
      required:
        - operating_system
        - architecture
        - format
        - primary_blob_digest
        - primary_blob_size_bytes
        - primary_media_type
      properties:
        operating_system:
          type: string
          pattern: '^[A-Za-z0-9][A-Za-z0-9._+-]{0,63}$'
        architecture:
          type: string
          pattern: '^[A-Za-z0-9][A-Za-z0-9._+-]{0,63}$'
        format:
          $ref: '#/components/schemas/ArtifactFormat'
        primary_blob_digest:
          $ref: '#/components/schemas/Digest'
        primary_blob_size_bytes:
          type: integer
          format: int64
          minimum: 0
        primary_media_type:
          type: string
          minLength: 1
      additionalProperties: false
    AddAttachmentRequest:
      type: object
      required:
        - name
        - media_type
        - blob_digest
        - blob_size_bytes
      properties:
        name:
          type: string
          pattern: '^[A-Za-z0-9][A-Za-z0-9._+-]{0,63}$'
        media_type:
          type: string
          minLength: 1
        blob_digest:
          $ref: '#/components/schemas/Digest'
        blob_size_bytes:
          type: integer
          format: int64
          minimum: 0
      additionalProperties: false
    Image:
      type: object
      required:
        - id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          pattern: '^[a-z0-9][a-z0-9._-]{0,127}$'
        display_name:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      additionalProperties: false
    ImageVersion:
      type: object
      required:
        - id
        - image_id
        - version
        - state
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        image_id:
          type: string
          format: uuid
        version:
          type: string
          pattern: '^[A-Za-z0-9][A-Za-z0-9._+:-]{0,127}$'
        state:
          $ref: '#/components/schemas/ImageVersionState'
        published_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      additionalProperties: false
    Artifact:
      type: object
      required:
        - id
        - version_id
        - operating_system
        - architecture
        - format
        - primary_blob_digest
        - primary_blob_size_bytes
        - primary_media_type
        - created_at
        - updated_at
      properties:
        id:
          $ref: '#/components/schemas/ArtifactID'
        version_id:
          type: string
          format: uuid
        operating_system:
          type: string
          pattern: '^[A-Za-z0-9][A-Za-z0-9._+-]{0,63}$'
        architecture:
          type: string
          pattern: '^[A-Za-z0-9][A-Za-z0-9._+-]{0,63}$'
        format:
          $ref: '#/components/schemas/ArtifactFormat'
        primary_blob_digest:
          $ref: '#/components/schemas/Digest'
        primary_blob_size_bytes:
          type: integer
          format: int64
          minimum: 0
        primary_media_type:
          type: string
          minLength: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      additionalProperties: false
    Attachment:
      type: object
      required:
        - id
        - artifact_id
        - name
        - media_type
        - blob_digest
        - blob_size_bytes
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        artifact_id:
          $ref: '#/components/schemas/ArtifactID'
        name:
          type: string
          pattern: '^[A-Za-z0-9][A-Za-z0-9._+-]{0,63}$'
        media_type:
          type: string
          minLength: 1
        blob_digest:
          $ref: '#/components/schemas/Digest'
        blob_size_bytes:
          type: integer
          format: int64
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      additionalProperties: false
    Manifest:
      type: object
      required:
        - image
        - version
        - artifacts
      properties:
        image:
          $ref: '#/components/schemas/Image'
        version:
          $ref: '#/components/schemas/ImageVersion'
        artifacts:
          type: array
          items:
            $ref: '#/components/schemas/ManifestArtifact'
      additionalProperties: false
    ManifestArtifact:
      type: object
      required:
        - artifact
        - attachments
      properties:
        artifact:
          $ref: '#/components/schemas/Artifact'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
      additionalProperties: false
    Problem:
      type: object
      description: >
        RFC 9457 problem details. The current API emits about:blank for all
        problem types; stable imgsrv-specific type URIs will be added only when
        clients need machine-readable branching beyond the HTTP status code.
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
          description: >
            RFC 9457 problem type. Clients should currently classify errors by
            HTTP status and must not parse detail.
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
      additionalProperties: true